% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_response_by_threshold.R
\name{add_response_by_threshold}
\alias{add_response_by_threshold}
\title{add_response_by_threshold}
\usage{
add_response_by_threshold(data = my_barrcoda_data, FC = 2, p_val = 0.001)
}
\arguments{
\item{my_baracoda_data}{baracoda_data}
}
\description{
A function to add response column "add_response_by_threshold"
}
