% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_data.R
\name{augment_data}
\alias{augment_data}
\title{augment data function}
\usage{
augment_data(data = my_clean_data)
}
\arguments{
\item{data}{data input}
}
\description{
This function takes in data and clean
}
