% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_func.R
\name{bar_plot_func}
\alias{bar_plot_func}
\title{Bar Plot Function}
\usage{
bar_plot_func(data = my_clean_augment_data, pep_length = 9)
}
\arguments{
\item{data}{data to plot}

\item{num}{Length of mut_peptide}
}
\description{
This function allows you to do a bar plot from specified peptide_position
}
