% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barc_resp.R
\name{barc_resp}
\alias{barc_resp}
\title{barcoda response generater}
\usage{
barc_resp(data = my_clean_augment_data, mouse_cell_line = "CT26")
}
\arguments{
\item{data}{data to use}

\item{mouse_cell_line}{select mouse cell line}
}
\description{
This function generates a sequence of logo.
}
