% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_frequency.R
\name{estimated_frequency}
\alias{estimated_frequency}
\title{estimated_frequency}
\usage{
estimated_frequency(data = my_barrcoda_data, fluochrome_dat = percent_FU)
}
\arguments{
\item{fluochrome_dat}{the importet percent flouchrome}

\item{my_baracoda_data}{barrcoda loaded data}
}
\description{
A function to calculate "estimated_frequency "
}
