% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploring_data_shiny.R
\name{Exploring_data_shiny}
\alias{Exploring_data_shiny}
\title{exploring data with shiny}
\usage{
Exploring_data_shiny(Plotting_data = my_clean_augment_data)
}
\arguments{
\item{Plotting_data}{data to plot}
}
\description{
This function takes in data and axplore in ashiny app
}
