% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load data function}
\usage{
load_data(data = baracoda_output)
}
\arguments{
\item{data}{log_fold_change file}
}
\description{
This function takes in excel fil and load
}
