% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_func.R
\docType{data}
\name{respond_cols}
\alias{respond_cols}
\title{Bar Plot Function}
\format{
An object of class \code{character} of length 2.
}
\usage{
respond_cols
}
\arguments{
\item{data}{data to plot}

\item{num}{Length of mut_peptide}
}
\description{
This function allows you to do a bar plot from specified peptide_position
}
\keyword{datasets}
