% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figures.R
\name{save_figures}
\alias{save_figures}
\alias{save_figure}
\title{save_figures}
\usage{
save_figure(plot = p, name = "figure", width = 12, height = 8)
}
\value{
A figure in results folder
}
\description{
A function to save "save_figure"
}
\examples{
save_figure(plot = p, name = "figure" , width = 12 , height = 8)
}
