% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_function.R
\name{scatterplot_function}
\alias{scatterplot_function}
\title{Scatter Plot Function}
\usage{
scatterplot_function(data, x, y)
}
\arguments{
\item{data}{data to plot}

\item{x}{x variable}

\item{y}{y variable}
}
\description{
This function takes in data, x, and y, and creates a scatter plot.
}
