% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqloggo_generator.R
\name{seqloggo_generator}
\alias{seqloggo_generator}
\title{Sequence Logo Generator}
\usage{
seqloggo_generator(
  data = x,
  len = y,
  resp = c("yes", "no"),
  cons = c("F", "M", "I", "D"),
  mouse = c("CT26", "4T1")
)
}
\description{
This function takes in a dataset and len.
}
