% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqloggo_generator.R
\name{augment_data}
\alias{augment_data}
\title{Augment the clean data}
\usage{
augment_data(
  data
)
}
\arguments{
\item{data}{clan data}

}
\description{
This function augment data and add estimated frequency from the baracoda files
}
