% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_func.R
\name{bar_plot_func}
\alias{bar_plot_func}
\title{Bar Plot Function}
\usage{
bar_plot_func(data = my_data_clean_aug, pep_length = 9, no_legend = TRUE)
}
\arguments{
\item{data}{data to plot}

\item{pep_length}{length of mut_peptide}

\item{no_length}{determine length}
}
\description{
This function allows you to do a bar plot from specified peptide_position
}
