% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barc_resp.R
\name{barc_resp}
\alias{barc_resp}
\title{barcoda response generater}
\usage{
barc_resp(data = my_barrcoda_data, est_freq = estimated_frequency)
}
\arguments{
\item{data}{data to use}

\item{est_freq}{the estimated frequency to use}
}
\description{
This function generates a sequence of logo.
}
