% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_function.R
\name{box_function}
\alias{box_function}
\title{Box Plot Function}
\usage{
box_function(data = data_single_peptides, x, y, no_legend = TRUE)
}
\arguments{
\item{data}{data to plot}

\item{x}{x variable}

\item{y}{y variable}

\item{no_legend}{determine legend}
}
\description{
This function takes in data, x, and y, and creates a box plot.
}
