% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_function.R
\name{box_function}
\alias{box_function}
\title{Box Plot Function}
\usage{
box_function(data, x, y)
}
\arguments{
\item{data}{data to plot}

\item{x}{x variable}

\item{y}{y variable}
}
\description{
This function takes in data, x, and y, and creates a box plot.
}
