% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqloggo_generator.R
\name{merge_all_data}
\alias{merge_all_data}
\title{Merge all data}
\usage{
merge_all_data(
  baracoda =  all_barracoda ,
                          mupexi  = all_mupexi,
                          sampel_information = sample_info,
                          buffycoat_HLA_information  = buffycoat.HLA_info
)
}

\arguments{
\item{baracoda}{all merge baracoda data}
\item{mupexi}{All merge mupexi files }
\item{sampel_information} {information about the input samples}
\item{buffycoat_HLA_information}{Buffycoaut HLA allele information}


}
\description{
This function cleans data
}
