% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqloggo_generator.R
\name{Exploring_data_shiny}
\alias{Exploring_data_shiny}
\title{Exploring_data_shiny}
\usage{
Exploring_data_shiny(Plotting_data = my_clean_augment_data
)
}

\arguments{
\item{Plotting_data}{data to plot , should be in the same format as "my_clean_augment_data""}

}
\description{
This function is to explore the data with a shiny app
}
