% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_FU_template_tab.R
\name{generate_FU_template_tab}
\alias{generate_FU_template_tab}
\title{generate_FU_template_tab}
\usage{
generate_FU_template_tab(
  data = my_barrcoda_data,
  outfile = "data/percent_PE.xlsx"
)
}
\arguments{
\item{data}{uploaded dataframe from my barracode data}
}
\description{
A function to generate template for PE percent table "generate_FU_template_tab"
}
