% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_all_data.R
\name{merge_all_data}
\alias{merge_all_data}
\title{merge data function}
\usage{
merge_all_data(
  all_barracoda = all_barracoda,
  all_mupexi = all_mupexi,
  sample_info = sample_info
)
}
\arguments{
\item{baracoda}{all merge baracoda files}

\item{mupexi}{all merge mupexi files}

\item{sampel_information}{all information about the sample in the data}

\item{buffycoat_HLA_information}{information about HLA in buffycoat}
}
\description{
This function takes in baracoda data, mupexi data, smaple information and buyyfyboat informatin
}
