% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_func.R, R/box_function.R
\docType{data}
\name{respond_cols}
\alias{respond_cols}
\title{Bar Plot Function}
\format{
An object of class \code{character} of length 2.

An object of class \code{character} of length 2.
}
\usage{
respond_cols

respond_cols
}
\arguments{
\item{num}{Length of mut_peptide}

\item{data}{data to plot}

\item{x}{x variable}

\item{y}{y variable}
}
\description{
This function allows you to do a bar plot from specified peptide_position

This function takes in data, x, and y, and creates a box plot.
}
\keyword{datasets}
