% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_function.R
\name{scatterplot_function}
\alias{scatterplot_function}
\title{Scatter Plot Function}
\usage{
scatterplot_function(
  data = my_clean_augment_data,
  x = "mut_mhcrank_el",
  y = "expression_level"
)
}
\arguments{
\item{data}{data to plot}

\item{x}{x variable}

\item{y}{y variable}
}
\description{
This function takes in data, x, and y, and creates a scatter plot.
}
