% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_function.R
\name{scatterplot_function}
\alias{scatterplot_function}
\title{Scatter Plot Function}
\usage{
scatterplot_function(
  data = data_single_peptides,
  x = "mut_mhcrank_el",
  y = "expression_level",
  no_legend = TRUE
)
}
\arguments{
\item{data}{data to plot}

\item{x}{x variable}

\item{y}{y variable}

\item{no_legend}{determind legend}
}
\description{
This function takes in data, x, and y, and creates a scatter plot.
}
