% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseq_logo.R
\name{seqloggo_generator}
\alias{seqloggo_generator}
\title{barcoda response generater}
\usage{
seqloggo_generator(data = my_data_clean_aug, len = 9, resp = c("yes", "no"))
}
\arguments{
\item{data}{data to use}

\item{len}{length of mut_peptide}

\item{resp}{response: yes or no}
}
\description{
This function generates a sequence of logo.
}
