% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqloggo_generator.R
\name{seqloggo_generator}
\alias{seqloggo_generator}
\title{Sequence Logo Generator}
\usage{
seqloggo_generator(
  data,
  len,
  resp = c("yes", "no"),
  cons = c("F", "M", "I", "D"),
  mouse = c("CT26", "4T1")
)
}
\arguments{
\item{data}{data to use}

\item{len}{length of mut_peptide}

\item{resp}{response: "yes" or "no}

\item{cons}{mutation consequence: "F", "M", "I", "D"}

\item{mouse}{cell_line of mouse}
}
\description{
This function generates a sequence of logo.
}
